+++ HiloNacional/src/main/java/com/HiloNacional/Repository/CompradorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompradorRepository extends JpaRepository<Comprador, Integer> {
+public interface CompradorRepository extends JpaRepository<Comprador, Long> {
 }
