+++ HiloNacional/src/main/java/com/HiloNacional/Repository/DetallePedidoRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Integer> {
+public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Long> {
 }
